LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/cross-origin-cookies-anonymous-client-ip-duplicate.https-expected.txt

CONSOLE MESSAGE: Prefetch request denied: not same origin as document
CONSOLE MESSAGE: Prefetch request denied: not same origin as document
CONSOLE MESSAGE: Prefetch request denied: not same origin as document
CONSOLE MESSAGE: Prefetch request denied: not same origin as document
CONSOLE MESSAGE: Prefetch request denied: not same origin as document
CONSOLE MESSAGE: Prefetch request denied: not same origin as document

FAIL speculation rules based prefetch should not use cookies for cross origin urls and should issue only the IP-anonymized request if that one is first. assert_equals: expected "prefetch" but got "navigate"
